Move YouTube's settings rows into our categories rather than doubling them
nativeSettings.js held the descent into the settings response, the rows we add, and
the rows we adopt, all at once. The shape is described once in settingsResponse.js
now, because both halves read it and both were descending separately — a title
arrives as runs on one renderer and as simpleText on another, and a comparison
against the wrong one silently never matches.

The rule adoptedRows.js exists for is not tidiness: where YouTube already ships a
switch for something, we move theirs instead of shipping a second one beside it. The
previews pair wrote the same ENABLE_PREVIEWS_WITH_SOUND flag from two places and ours
re-forced it on every page load, so turning YouTube's off did not stay off.
