+++ mods/settings/nativeSettings.js
@@ -1,230 +1,41 @@
-import { configChangeEmitter, configRead } from '../../framework/index.js';
-import { GROUPS, chosenLabel, chosenSummary } from './settingsModel.js';
-import { optionsCommand, storeCommand } from './settingsOptions.js';
-import { claimBooleanRows, claimActionRows, redrawSettingRows, ROWS } from './settingComponents.js';
-import { claimVersionPanel } from './aboutPatch.js';
+import { configChangeEmitter, onResponse } from '../../framework/index.js';
+import { GROUPS } from './settingsModel.js';
+import { categoryFor } from './settingRows.js';
+import { ADDED, putMoved, takeMoved } from './adoptedRows.js';
+import { category, categoryOf, indexOfCategory } from './settingsResponse.js';
+import { claimDrawnRows } from './claimDrawn.js';
+import { redrawSettingRows } from './settingComponents.js';
+
+// Our settings, in YouTube's own settings page.
+//
+// Three things happen to the response, in this order, and the order is the point: our categories
+// go in first so there is somewhere for YouTube's rows to land, its rows are then taken out of
+// wherever it put them, and the categories that exist only to hold them are added last — only if
+// anything landed.
 
 const BEFORE = 'SETTING_CAT_TVHTML5_LINK_PHONE';
 
 const ABOUT = 'SETTING_CAT_TVHTML5_ABOUT';
 
-const runs = (text) => ({ runs: [{ text }] });
-const picture = (url) => ({ thumbnails: [{ url }] });
-const idFor = (item) => `TUBE_${item.key || item.id}`;
+const insertAt = (items, index, added) => items.splice.apply(items, [index, 0].concat(added));
 
-const categoryOf = (item) => item && item.settingCategoryCollectionRenderer;
-
-function switchRow(item) {
-    return {
-        settingBooleanRenderer: {
-            title: runs(item.title),
-            summary: runs(item.summary),
-            enabled: configRead(item.key) === item.on,
-            enableServiceEndpoint: storeCommand(item.key, item.on),
-            disableServiceEndpoint: storeCommand(item.key, !item.on),
-            trackingParams: 'null',
-            itemId: idFor(item),
-            thumbnail: picture(item.image)
-        }
-    };
-}
-
-function openerRow(item, path, label) {
-    const row = {
-        title: runs(item.title),
-        summary: runs(item.summary),
-        serviceEndpoint: optionsCommand(path),
-        tubeNote: item.note,
-        trackingParams: 'null',
-        itemId: idFor(item),
-        thumbnail: picture(item.image)
-    };
-
-    Object.defineProperty(row, 'actionLabel', {
-        enumerable: true,
-        get: () => runs(label())
-    });
-
-    return { settingActionRenderer: row };
-}
-
-const rowFor = (item, path) =>
-    item.kind === 'switch' ? switchRow(item)
-        : item.kind === 'choice'
-            ? openerRow(item, path, () => `${item.prefix}: ${chosenLabel(item)}`)
-            : openerRow(item, path, () => chosenSummary(item));
-
-const category = (categoryId, title, items) => ({
-    settingCategoryCollectionRenderer: {
-        categoryId,
-        title: runs(title),
-        focused: false,
-        trackingParams: 'null',
-        items
-    }
+// Every settings response is patched, and the guide is fetched more than once per session. Ours
+// being there already is how a second pass is recognised: patching twice would double every row.
+const alreadyPatched = (items) => items.some((item) => {
+    const found = categoryOf(item);
+    return found && GROUPS.some((group) => group.id === found.categoryId);
 });
-
-const categoryFor = (group, groupIndex) => category(
-    group.id,
-    group.title,
-    group.items.map((item, index) => rowFor(item, [groupIndex, index]))
-);
-
-function indexOfCategory(items, categoryId) {
-    const index = items.findIndex((item) => {
-        const found = categoryOf(item);
-        return found && found.categoryId === categoryId;
-    });
-
-    return index === -1 ? items.length : index;
-}
-
-function findCategory(items, categoryId) {
-    const index = indexOfCategory(items, categoryId);
-    return index === items.length ? null : categoryOf(items[index]);
-}
-
-const SUBSCRIPTION = 'tube_subscription';
-const PARENTAL = 'tube_parental';
-const HISTORY = 'SETTING_CAT_TVHTML5_HISTORY';
-const PLAYBACK = 'tube_playback';
-const INTERFACE = 'tube_interface';
-
-const MOVES = [
-    { to: SUBSCRIPTION, id: 'PREMIUM_LANDING_PAGE', title: 'Get YouTube Premium' },
-    { to: SUBSCRIPTION, id: 'MANAGE_PURCHASES_AND_MEMBERSHIPS', title: 'Purchases and memberships' },
-    { to: PLAYBACK, id: 'AUTONAV', title: 'Autoplay next video' },
-    { to: PARENTAL, setting: 'SAFETY_MODE', title: 'Restricted mode' },
-    { to: PARENTAL, id: 'PARENT_CODE', title: 'Parent code' },
-    { to: HISTORY, id: 'RECOMMEND', title: 'Device recommendations', first: true },
-    { to: INTERFACE, setting: 'ENABLE_PREVIEWS_WITH_SOUND', title: 'Previews' }
-];
-
-const ADDED = [
-    { id: SUBSCRIPTION, title: 'Subscription' },
-    { id: PARENTAL, title: 'Parental controls' }
-];
-
-const textOf = (text) => {
-    if (!text) return '';
-    if (Array.isArray(text.runs)) return text.runs.map((run) => run.text).join('');
-    return text.simpleText || '';
-};
-
-const rendererOf = (row) => {
-    const key = row && Object.keys(row)[0];
-    return key ? row[key] : null;
-};
-
-const clientSettingOf = (renderer) => {
-    const endpoint = renderer.enableServiceEndpoint || renderer.serviceEndpoint;
-    const datas = endpoint
-        && endpoint.setClientSettingEndpoint
-        && endpoint.setClientSettingEndpoint.settingDatas;
-
-    const chosen = datas && datas[0] && datas[0].clientSettingEnum;
-    return chosen ? chosen.item : '';
-};
-
-function moveFor(row, reachable) {
-    const renderer = rendererOf(row);
-    if (!renderer) return null;
-
-    const id = renderer.itemId || '';
-    const setting = clientSettingOf(renderer);
-    const title = textOf(renderer.title);
-
-    const move = MOVES.find((rule) =>
-        (rule.id && id.indexOf(rule.id) !== -1)
-        || (rule.setting && rule.setting === setting)
-        || rule.title === title);
-
-    return move && reachable[move.to] ? move : null;
-}
-
-function takeMoved(items) {
-    const reachable = {};
-    ADDED.forEach((entry) => { reachable[entry.id] = true; });
-    items.forEach((item) => {
-        const found = categoryOf(item);
-        if (found) reachable[found.categoryId] = true;
-    });
-
-    const categories = items.map(categoryOf).filter((found) => found && Array.isArray(found.items));
-
-    const moved = categories.reduce((all, found) => all.concat(found.items
-        .map((row) => ({ row, move: moveFor(row, reachable) }))
-        .filter((entry) => entry.move)), []);
-
-    const taken = moved.reduce((byTarget, entry) => ({
-        ...byTarget,
-        [entry.move.to]: (byTarget[entry.move.to] || []).concat([entry])
-    }), {});
-
-    categories.forEach((found) => {
-        found.items = found.items.filter((row) => !moveFor(row, reachable));
-    });
-
-    // Emptied categories are removed after the walk, not during it: splicing mid-iteration skips
-    // whatever followed each removal.
-    const emptied = items.filter((item) => {
-        const found = categoryOf(item);
-        return found && Array.isArray(found.items) && found.items.length === 0;
-    });
-
-    emptied.forEach((item) => items.splice(items.indexOf(item), 1));
-
-    return taken;
-}
-
-function putMoved(items, taken) {
-    Object.keys(taken).forEach((categoryId) => {
-        const found = findCategory(items, categoryId);
-        if (!found) return;
-
-        const rows = taken[categoryId];
-        const lead = rows.filter((entry) => entry.move.first).map((entry) => entry.row);
-        const rest = rows.filter((entry) => !entry.move.first).map((entry) => entry.row);
-
-        found.items = lead.concat(found.items, rest);
-    });
-}
-
-const SETTLING_ATTEMPTS = 20;
-const PATIENT_ATTEMPTS = 120;
-const SETTLING_EVERY = 250;
-const PATIENT_EVERY = 500;
-
-// The action row's component loads only when the first row is drawn, after this response is patched.
-function claimRows(attempt = 0) {
-    const claimed = claimBooleanRows();
-    const annotated = claimActionRows();
-    const stamped = claimVersionPanel();
-    if (claimed && annotated && stamped) return;
-
-    const settling = attempt < SETTLING_ATTEMPTS;
-    if (!settling && !document.querySelector(ROWS)) return;
-    if (attempt >= SETTLING_ATTEMPTS + PATIENT_ATTEMPTS) return;
-
-    setTimeout(() => claimRows(attempt + 1), settling ? SETTLING_EVERY : PATIENT_EVERY);
-}
 
 function PatchSettings(response) {
     if (!Array.isArray(response.items)) return;
 
     const items = response.items;
     if (!items.some(categoryOf)) return;
+    if (alreadyPatched(items)) return;
 
-    const already = (item) => {
-        const found = categoryOf(item);
-        return found && GROUPS.some((group) => group.id === found.categoryId);
-    };
-    if (items.some(already)) return;
+    claimDrawnRows();
 
-    claimRows();
-
-    const categories = GROUPS.map(categoryFor);
-    items.splice.apply(items, [indexOfCategory(items, BEFORE), 0].concat(categories));
+    insertAt(items, indexOfCategory(items, BEFORE), GROUPS.map(categoryFor));
 
     const taken = takeMoved(items);
 
@@ -232,10 +43,14 @@ function PatchSettings(response) {
         .filter((entry) => taken[entry.id] && taken[entry.id].length)
         .map((entry) => category(entry.id, entry.title, []));
 
-    items.splice.apply(items, [indexOfCategory(items, ABOUT), 0].concat(added));
+    insertAt(items, indexOfCategory(items, ABOUT), added);
 
     putMoved(items, taken);
 }
+
+// Registered here rather than called from the adblocker, which is where the whole settings UI
+// used to be loaded from as a side effect.
+onResponse('settings', ['items'], PatchSettings);
 
 const start = () => {
     configChangeEmitter.addEventListener('configChange', () => redrawSettingRows());
